+++ 3rdParty/LightPcapNg/LightPcapNg/src/light_pcapng.c
@@ -303,7 +303,7 @@ void light_read_record(light_file fd, light_pcapng *record)
    //See the block type, if end of file this will tell us
    uint32_t blockType, blockSize, bytesRead;
    bytesRead = light_read(fd, &blockType, sizeof(blockType));
-   if (bytesRead != sizeof(blockType) || (bytesRead == EOF && feof(fd->file)))
+   if (bytesRead != sizeof(blockType) || (bytesRead == (uint32_t)EOF && feof(fd->file)))
    {
       current = NULL;
       return;
@@ -318,7 +318,7 @@ void light_read_record(light_file fd, light_pcapng *record)
 
    //Get block size
    bytesRead = light_read(fd, &current->block_total_length, sizeof(blockSize));
-   if (bytesRead != sizeof(blockSize) || (bytesRead == EOF && feof(fd->file)))
+   if (bytesRead != sizeof(blockSize) || (bytesRead == (uint32_t)EOF && feof(fd->file)))
    {
       free(current);
       current = NULL;
@@ -332,7 +332,7 @@ void light_read_record(light_file fd, light_pcapng *record)
    const uint32_t bytesToRead = current->block_total_length - 2 * sizeof(blockSize) - sizeof(blockType);
    uint32_t *local_data = calloc(bytesToRead, 1);
    bytesRead = light_read(fd, local_data, bytesToRead);
-   if (bytesRead != bytesToRead || (bytesRead == EOF && feof(fd->file)))
+   if (bytesRead != bytesToRead || (bytesRead == (uint32_t)EOF && feof(fd->file)))
    {
       free(current);
       free(local_data);
@@ -343,7 +343,7 @@ void light_read_record(light_file fd, light_pcapng *record)
    //Need to move file to next record so read the footer, which is just the record length repeated
    bytesRead = light_read(fd, &blockSize, sizeof(blockSize));
    //Verify the two sizes match!!
-   if (blockSize != current->block_total_length || bytesRead != sizeof(blockSize) || (bytesRead == EOF && feof(fd->file)))
+   if (blockSize != current->block_total_length || bytesRead != sizeof(blockSize) || (bytesRead == (uint32_t)EOF && feof(fd->file)))
    {
       free(current);
       free(local_data);
